nopoll: * [fix] Some changes to compile regression-test on windows7 (64 bits).

+++ trunk/test/nopoll-regression-client.c
@@ -299,7 +299,7 @@ nopoll_bool test_01 (void) {
 
 		if (! nopoll_conn_is_ok (conn)) {
 			printf ("ERROR (4.1 jkd412): expected to find proper connection handshake finished, but found connection is broken: session=%d, errno=%d : %s..\n",
-				nopoll_conn_socket (conn), errno, strerror (errno));
+				(int) nopoll_conn_socket (conn), errno, strerror (errno));
 			return nopoll_false;
 		} /* end if */
 
@@ -967,7 +967,7 @@ nopoll_bool test_06 (void) {
 
 		if (! nopoll_conn_is_ok (conn)) {
 			printf ("ERROR (4.1 jg72): expected to find proper connection handshake finished, but found connection is broken: session=%d, errno=%d : %s..\n",
-				nopoll_conn_socket (conn), errno, strerror (errno));
+				(int) nopoll_conn_socket (conn), errno, strerror (errno));
 			return nopoll_false;
 		} /* end if */
 
@@ -1015,7 +1015,7 @@ nopoll_bool test_07 (void) {
 
 		if (! nopoll_conn_is_ok (conn)) {
 			printf ("ERROR (4.1 dk45): expected to find proper connection handshake finished, but found connection is broken: session=%d, errno=%d : %s..\n",
-				nopoll_conn_socket (conn), errno, strerror (errno));
+				(int) nopoll_conn_socket (conn), errno, strerror (errno));
 			return nopoll_false;
 		} /* end if */
